Added memory checking to all tests. Added test to ensure memory safty in the case of a bad read.

+++ test/test.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 #
-# A script to run tests. Takes three arguments
+# A script to run tests. Takes four arguments
 #
 # 1) the binary or script to be run
 # 2) the expected exit status
@@ -39,6 +39,18 @@ if [ ${DIFF_EXIT_STAT} != 0 ] ; then
   echo "Diff of outputs uncovered differences. [ FAIL ]"
   exit 1
 fi
+
+########## MEMORY CHECK
+
+NOLEAKS=$(valgrind ${BIN} 2>&1 | grep "no leaks are possible" | wc -l)
+if [ x${NOLEAKS} == x1 ] ; then
+  echo "Test assertions all held, but a memory leak was detected."
+  echo "For more details run: valgrind ${BIN}"
+  echo "[ FAIL ]"
+  exit 1
+fi
+
+
 
 rm -f ${TMP_OUT_FILE}
